Fix remaining cross-platform test build issues
Fixed three platform-specific build issues:

1. macOS: Changed brew install from 'obs-studio' to '--cask obs'
   - OBS Studio is distributed as a cask on macOS, not a formula

2. Ubuntu/Linux: Added obs_stubs.c to provide mock obs_log function
   - Source code uses obs_log() but tests don't need full OBS logger
   - Created stub implementation that suppresses logging in tests
   - Resolves "undefined reference to obs_log" linking errors

3. Windows: Fixed type conversion warning in mock_restreamer.c
   - Added explicit cast: send(..., (int)strlen(response), ...)
   - Windows treats warnings as errors, needed explicit size_t->int cast

All platforms should now build and link successfully.

+++ tests/mock_restreamer.c
@@ -113,7 +113,7 @@ static void handle_request(int client_fd, const char *request) {
   }
 
   /* Send response */
-  send(client_fd, response, strlen(response), 0);
+  send(client_fd, response, (int)strlen(response), 0);
 }
 
 /* Server thread function */
